third_party/blink/web_tests/http/tests/inspector-protocol/animation/animation-scroll-timeline-updated-event-expected.txt

Tests that `animationUpdated` event is emitted when the animation is updated
PASS: startTime is reported correctly as percentage.
PASS: startTime is reported correctly as percentage in the animation coming from animationUpdated event.
